Simplified components for showing file with search functionality and getDocuments; Renamed DocumentModals directory to Modals
