--- src/components/DocumentModals/index.js
@@ -1,11 +1,0 @@
-import ShowDocumentPreviewModal from './ShowDocumentPreviewModal';
-import ShowDocumentsModal from './ShowDocumentsModal';
-
-/**
- * Components and functions related to document viewer functionality within
- * project PASS
- *
- * @namespace DocumentModals
- */
-
-export { ShowDocumentPreviewModal, ShowDocumentsModal };
