+++ src/routes/index.js
@@ -1,5 +1,4 @@
 import Clients from './Clients';
-import User from './User';
 import Forms from './Forms';
 import Home from './Home';
 import Inbox from './Inbox';
@@ -11,4 +10,4 @@ import Outbox from './Outbox';
  * @namespace Pages
  */
 
-export { Clients, User, Forms, Home, Inbox, Outbox };
+export { Clients, Forms, Home, Inbox, Outbox };
